final fix + cleanup
