+++ .latch/workflow_name
@@ -1,1 +1,1 @@
-wf.entrypoint.nf_nf_core_methylseq
+nf_nf_core_methylseq
