+++ version
@@ -1,1 +1,1 @@
-0.3.0
+2.6.0
